refactor: move Markets component into own folder
